Mise en page sous forme de tableau #76

+++ source/achete_ta_baguette_fr_administrateur/action/action-produit.php
@@ -23,55 +23,70 @@ $page = (object)
 ];
 
 
-
 $laBDD = new AccesseurProduit();
 $laBDD2 = new AccesseurCategorie();
-$page->listeProduits = $laBDD->recupererListeProduits();
-$page->listeCategorie = $laBDD2->recupererlisteCategorie();
 
-function supprimerProduit ($produit){
-	$laBDD->supprimerProduit($produit);
+
+function modifierProduit ($produit){
+	echo 'test';
+	$laBDD->modifierProduit($produit);
+	
 }
 function recupererLabelCategorieParProduit($page, $produit){
 	
-		foreach ($page->listeCategorie as $categorie) {
-			if($categorie->getIdCategorie() == $produit->getIdCategorie()) return $categorie->getLabel();
-
-		}
-
+	foreach ($page->listeCategorie as $categorie) {
+		if($categorie->getIdCategorie() == $produit->getIdCategorie()) return $categorie->getLabel();
 
 	}
 
-	$laRedirection = new Redirection("/administration/vue/index.php");
 
-	if(isset($_POST['action-modifier-produit'])){ 
+}
 
-		$attribut = (object)
-		[
-			"nom" => "",
-			"categorie" => "",
-			"description" => "",
-			"stock" => null,
-			"prix" => "12",
-			"srcImage" => "test.png", 
-		];
-		$produit = new Produit($attribut);
-		$produit->setNom($_POST['produit']);
-		$produit->setIdCategorie($_POST['categorie']);
-		$produit->setDescription($_POST['description']);
-		$produit->setStock($_POST['stock']);
-		$produit->setPrix($_POST['prix']);
-		$laBDD->ajouterProduit($produit);
-		$page->addProduit = true;
+$laRedirection = new Redirection("/administration/vue/index.php");
 
-		if($page->addProduit ?? false){
+if(isset($_POST['action-ajouter-produit'])){ 
+
+	$attribut = (object)
+	[
+		"nom" => "",
+		"categorie" => "",
+		"description" => "",
+		"stock" => null,
+		"prix" => "12",
+		"srcImage" => "test.png", 
+	];
+	$produit = new Produit($attribut);
+	$produit->setNom($_POST['produit']);
+	$produit->setIdCategorie($_POST['categorie']);
+	$produit->setDescription($_POST['description']);
+	$produit->setStock($_POST['stock']);
+	$produit->setPrix($_POST['prix']);
+	$laBDD->ajouterProduit($produit);
+	$page->addProduit = true;
+
+		/*if($page->addProduit ?? false){
 			header("Location: " . "/administration/vue/index.php");
 			exit;
 			$page->addProduit = false;
-		}
+		}*/
 
-	}
+}
 
-	$page->listeProduits = $laBDD->recupererListeProduits();
-	$page->listeCategorie = $laBDD2->recupererlisteCategorie();
-	?>
+if(isset($_POST['action-supprimer-produit'])){ 
+	
+	$produit = new Produit((object)$_POST);
+	$laBDD->supprimerProduit($produit);	
+
+}
+
+if(isset($_POST['action-modifier-produit'])){
+
+	$produit = new Produit((object)$_POST);
+	$laBDD->modifierProduit($produit);
+
+}
+
+$page->listeProduits = $laBDD->recupererListeProduits();
+$page->listeCategorie = $laBDD2->recupererListeCategorie();
+
+?>
